Particle Updates
Disk particles are more round

There is a firetrail emitter that activates on death

There are 2 emitters on the ufo's sides that make a cool light trail effect

The explosion is a little higher so it is in view

Particle was updated to have rotations and a matrix

There is now a cone shaped emitter which was used for firetrail

The emitter can now take an array of colors to randomize from. The emitter also sets particle rotation acceleration now.

The timer starts at 0 so as soon as an emitter is activated it emits.